requirements.txt
langchain
faiss-cpu
requests
pytest
tqdm
sentence-transformers
numpy
transformers
beautifulsoup4
